+++ Changes.md
@@ -23,6 +23,11 @@ Build
 
 - Cortex : Updated to version 10.6.1.0.
 
+API
+---
+
+- PopupWindow : Added `showWarning()` class method.
+
 1.6.3.0 (relative to 1.6.2.1)
 =======